services/agent/requirements.txt
paho-mqtt==2.1.0
requests==2.32.3
influxdb-client==1.48.0
